+++ ElvUI_LuckyoneUI/Profiles/AddOnSkins.lua
@@ -1,7 +1,11 @@
-local _, Private = ...
-local E, L = unpack(ElvUI)
-
+-- Lua functions
 local unpack = unpack
+
+-- AddOn namespace
+local _, Private = ...
+
+-- ElvUI modules
+local E, L = unpack(ElvUI)
 
 -- AddOnSkins profile
 function Private:Setup_AddOnSkins()