+++ ElvUI_LuckyoneUI/Settings/Defaults.lua
@@ -1,4 +1,10 @@
+-- Lua functions
+local unpack = unpack
+
+-- AddOn namespace
 local _, Private = ...
+
+-- ElvUI modules
 local _, _, V, P, G = unpack(ElvUI)
 
 -- Defaults: E.global.L1UI