+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+### Version 3.03 [24-07-2024]
+**All WoW Versions:**  
+ElvUI 13.71+ compatibility  
+Increased max possible rows of buffs for target and focus from 2 to 3  
+The quick setup for alts now includes the custom chat setup  
+**Classic:**  
+Enabled mana mp5 ticks by default in the healing layout  
+Fixed pet name tag to correctly work with temporary pets like Shadowfiend  
+Updated text format for the player mana bar in the healing layout  
+**Retail:**  
+Updated TOC for 11.0  
+___
 ### Version 3.02 [10-07-2024]
 **All WoW Versions:**  
 Added quick setup for alts which can load existing LuckyoneUI profiles